REF/TST: misplaced tests in test_timeseries, test_timezones (#32300)
