+++ pandas/tests/frame/test_constructors.py
@@ -7,6 +7,7 @@ import numpy as np
 import numpy.ma as ma
 import numpy.ma.mrecords as mrecords
 import pytest
+import pytz
 
 from pandas.compat import is_platform_little_endian
 from pandas.compat.numpy import _is_numpy_dev
@@ -2389,6 +2390,12 @@ class TestDataFrameConstructors:
         result = DataFrame.from_records(data)
         tm.assert_frame_equal(result, expected)
 
+    def test_frame_from_records_utc(self):
+        rec = {"datum": 1.5, "begin_time": datetime(2006, 4, 27, tzinfo=pytz.utc)}
+
+        # it works
+        DataFrame.from_records([rec], index="begin_time")
+
     def test_to_frame_with_falsey_names(self):
         # GH 16114
         result = Series(name=0, dtype=object).to_frame().dtypes
@@ -2459,6 +2466,18 @@ class TestDataFrameConstructors:
             columns=["f", "female", "m", "male", "unknown"],
         )
         tm.assert_frame_equal(result, expected)
+
+    def test_from_M8_structured(self):
+        dates = [(datetime(2012, 9, 9, 0, 0), datetime(2012, 9, 8, 15, 10))]
+        arr = np.array(dates, dtype=[("Date", "M8[us]"), ("Forecasting", "M8[us]")])
+        df = DataFrame(arr)
+
+        assert df["Date"][0] == dates[0][0]
+        assert df["Forecasting"][0] == dates[0][1]
+
+        s = Series(arr["Date"])
+        assert isinstance(s[0], Timestamp)
+        assert s[0] == dates[0][0]
 
 
 class TestDataFrameConstructorWithDatetimeTZ:
